requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
ollama
pandas
alembic
pytest
httpx
